+++ tests/TestDataService/TestDataService.spec.ts
@@ -19,6 +19,7 @@ import {
     type User,
     type AclRole,
     type CmsPage,
+    type SalesChannel,
 } from "../../src";
 
 test("Data Service", async ({ TestDataService, AdminApiContext }) => {
@@ -44,6 +45,16 @@ test("Data Service", async ({ TestDataService, AdminApiContext }) => {
     const currency = await TestDataService.createCurrency({ taxFreeFrom: 10 });
     expect(currency.taxFreeFrom).toEqual(10);
 
+    const salesChannelDomain = await TestDataService.createSalesChannelDomain();
+    const salesChannelResponse = await AdminApiContext.post("./search/sales-channel", {
+        data: {
+            filter: [{ type: "equals", field: "id", value: TestDataService.defaultSalesChannel.id }],
+            associations: { currencies: {} },
+        },
+    });
+    const { data: salesChannels } = (await salesChannelResponse.json()) as { data: SalesChannel[] };
+    expect(salesChannels[0].currencies).toContainEqual(expect.objectContaining({ id: salesChannelDomain.currencyId }));
+
     const country = await TestDataService.createCountry();
     expect(country.name).toBeDefined();
 
@@ -59,6 +70,7 @@ test("Data Service", async ({ TestDataService, AdminApiContext }) => {
 
     const digitalProduct = await TestDataService.createDigitalProduct("Test Test", { description: "You can download me." });
     expect(digitalProduct.description).toEqual("You can download me.");
+    expect(digitalProduct.states).toContain("is-download");
 
     const propertyGroup = await TestDataService.createColorPropertyGroup();
     expect(propertyGroup.description).toEqual("Color");